backend/requirements.txt
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
SQLAlchemy==2.0.30
alembic==1.12.0
celery==5.3.0
redis==4.6.0
httpx==0.27.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
openai==1.11.0
python-dotenv==1.0.0
pydantic-settings==2.13.1
psycopg[binary]==3.2.10
pytest==8.3.5
